requirements.txt
PyPDF2
langchain
langchain_community
langchain_text_splitters
langchain_groq
flask
python-dotenv